# Conflicts:
#	Assets/Prefabs/Levels/Level1.prefab
#	Assets/Scripts/Character.cs
#	Assets/SpineAnimations.meta
#	Assets/SpineAnimations/CatAnimations.meta
#	Assets/SpineAnimations/CatAnimations/ReferenceAssets.meta
#	Assets/SpineAnimations/RabbitAnimmations.meta

+++ Assets/Scripts/GameManager.cs
@@ -11,7 +11,7 @@ public class GameManager : MonoBehaviour
     public GameMode gameMode = GameMode.StoryMode;
     public NetworkMode networkMode = NetworkMode.Local;
 
-    public int requiredRootsAmount;
+    public float requiredRootsAmount;
     private int collectedRootsAmount;
 
     public static UnityAction OncollectedCounterUpdated;
